fix bmp, image

+++ solution/include/image.h
@@ -2,6 +2,8 @@
 #define IMAGE_H
 
 #include <stdint.h>
+#include <stdio.h>
+#include <stdlib.h>
 
 // Структура для внутреннего представления картинки
 struct pixel {
@@ -12,14 +14,14 @@ struct pixel {
 };
 
 struct image {
-    uint64_t width;
-    uint64_t height;
+    uint32_t width;
+    uint32_t height;
     struct pixel* data;
     uint8_t* padding;
 };
 
 // Функция для создания изображения
-struct image* create_image(uint64_t width, uint64_t height);
+struct image* create_image(uint32_t width, uint32_t height);
 
 // Теперь Добби свободен
 void free_image(struct image* img);
